v1/requirements.txt
Flask==2.1.1
PyYAML==6.0
ndjson==0.3.1
google-cloud-bigquery==3.0.1
google-cloud-storage==2.3.0
google-cloud-pubsub==2.12.0